Refine toolchain header handling
The ninja file contains all the necessary info about
include search paths. It only ruins the search process
to explicitly add the include path from the toolchain.

To precisely determine if a header comes from toolchain,
it needs more info such as some keywords in the toolchain
path. That will add another parameter. And the benefit is
not so much. So do not tell if a header comes from toolchain
dir for now. Maybe later I can find a better way.

Signed-off-by: Ming Shao <smrtos@163.com>